move repository interface to domain
