+++ app/src/main/java/ru/practicum/android/diploma/di/RepositoryModule.kt
@@ -1,7 +1,7 @@
 package ru.practicum.android.diploma.di
 
 import org.koin.dsl.module
-import ru.practicum.android.diploma.data.search.SearchRepository
+import ru.practicum.android.diploma.domain.repository.SearchRepository
 import ru.practicum.android.diploma.data.search.SearchRepositoryImpl
 
 val repositoryModule = module {
