+++ app/src/main/java/ru/practicum/android/diploma/domain/repository/SearchRepository.kt
@@ -1,4 +1,4 @@
-package ru.practicum.android.diploma.data.search
+package ru.practicum.android.diploma.domain.repository
 
 import kotlinx.coroutines.flow.Flow
 import ru.practicum.android.diploma.domain.models.VacancyList
